+++ zetaclient/broadcast.go
@@ -14,6 +14,7 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	flag "github.com/spf13/pflag"
 	rpchttp "github.com/tendermint/tendermint/rpc/client/http"
+	"github.com/zeta-chain/zetacore/app/ante"
 	"github.com/zeta-chain/zetacore/cmd/zetacored/config"
 	"github.com/zeta-chain/zetacore/common/cosmos"
 	"github.com/zeta-chain/zetacore/zetaclient/hsm"
@@ -41,6 +42,9 @@ func (b *ZetaCoreBridge) Broadcast(gaslimit uint64, authzWrappedMsg sdktypes.Msg
 	if baseGasPrice == 0 {
 		baseGasPrice = DefaultBaseGasPrice // shoudn't happen, but just in case
 	}
+	reductionRate := sdktypes.MustNewDecFromStr(ante.GasPriceReductionRate)
+	// multiply gas price by the system tx reduction rate
+	adjustedBaseGasPrice := sdktypes.NewDec(baseGasPrice).Mul(reductionRate)
 
 	if blockHeight > b.blockHeight {
 		b.blockHeight = blockHeight
@@ -72,7 +76,7 @@ func (b *ZetaCoreBridge) Broadcast(gaslimit uint64, authzWrappedMsg sdktypes.Msg
 	builder.SetGasLimit(gaslimit)
 	// #nosec G701 always in range
 	fee := sdktypes.NewCoins(sdktypes.NewCoin(config.BaseDenom,
-		cosmos.NewInt(int64(gaslimit)).Mul(cosmos.NewInt(baseGasPrice))))
+		cosmos.NewInt(int64(gaslimit)).Mul(adjustedBaseGasPrice.Ceil().RoundInt())))
 	builder.SetFeeAmount(fee)
 	//fmt.Printf("signing from name: %s\n", ctx.GetFromName())
 	err = b.SignTx(factory, ctx.GetFromName(), builder, true, ctx.TxConfig)