+++ zetaclient/zetacore_observer.go
@@ -2,8 +2,11 @@ package zetaclient
 
 import (
 	"fmt"
+	"math"
 	"strings"
 	"time"
+
+	observertypes "github.com/zeta-chain/zetacore/x/observer/types"
 
 	sdkmath "cosmossdk.io/math"
 
@@ -158,7 +161,7 @@ func (co *CoreObserver) startCctxScheduler() {
 						co.logger.ZetaChainWatcher.Error().Err(err).Msgf("couldn't get operator balance")
 					} else {
 						diff := co.lastOperatorBalance.Sub(balance)
-						if diff.GT(sdkmath.NewInt(0)) {
+						if diff.GT(sdkmath.NewInt(0)) && diff.LT(sdkmath.NewInt(math.MaxInt64)) {
 							co.ts.AddFeeEntry(bn, diff.Int64())
 							co.lastOperatorBalance = balance
 						}
@@ -355,7 +358,7 @@ func (co *CoreObserver) getUpdatedChainOb(chainID int64) (ChainClient, error) {
 	curParams := chainOb.GetChainParams()
 	if common.IsEVMChain(chainID) {
 		evmCfg, found := co.cfg.GetEVMConfig(chainID)
-		if found && curParams != evmCfg.ChainParams {
+		if found && !observertypes.ChainParamsEqual(curParams, evmCfg.ChainParams) {
 			chainOb.SetChainParams(evmCfg.ChainParams)
 			co.logger.ZetaChainWatcher.Info().Msgf(
 				"updated chain params for chainID %d, new params: %v",
@@ -365,7 +368,7 @@ func (co *CoreObserver) getUpdatedChainOb(chainID int64) (ChainClient, error) {
 		}
 	} else if common.IsBitcoinChain(chainID) {
 		_, btcCfg, found := co.cfg.GetBTCConfig()
-		if found && curParams != btcCfg.ChainParams {
+		if found && !observertypes.ChainParamsEqual(curParams, btcCfg.ChainParams) {
 			chainOb.SetChainParams(btcCfg.ChainParams)
 			co.logger.ZetaChainWatcher.Info().Msgf(
 				"updated chain params for Bitcoin, new params: %v",