+++ changelog.md
@@ -13,6 +13,10 @@
 * [1535](https://github.com/zeta-chain/node/issues/1535) - Avoid voting on wrong ballots due to false blockNumber in EVM tx receipt
 * [1588](https://github.com/zeta-chain/node/pull/1588) - fix chain params comparison logic
 
+### CI
+
+* [1580](https://github.com/zeta-chain/node/pull/1580) - Fix release pipelines cleanup step.
+
 ### Chores
 
 * [1585](https://github.com/zeta-chain/node/pull/1585) - Updated release instructions